requirements-test.txt
pytest
fastapi
uvicorn
sqlalchemy
pydantic
Flask
Flask-JWT-Extended
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv
